backend/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.3.1
gunicorn==21.2.0
mysqlclient==2.2.0
psycopg2-binary==2.9.9
django-rest-passwordreset>=1.3.0
python-decouple==3.8
django-storages[boto3]==1.14.2
whitenoise==6.6.0

# Code Quality Tools
black==23.12.1
flake8==7.0.0
isort==5.13.2
pylint==3.0.3
python-dotenv
pytest==7.4.2
pytest-django
pytest-cov

# WebSocket / Channels
channels==4.0.0
channels-redis==4.1.0
daphne==4.0.0
